Refactored out a superclass for the different job types. Also added some tests of __str__ and __repr__ for the different job types.
