+++ src/relion/_parser/class3D.py
@@ -1,9 +1,6 @@
-import collections.abc
-from gemmi import cif
-import os
-import functools
 from collections import namedtuple
 from collections import Counter
+from relion._parser.jobtype import JobType
 
 
 Class3DParticleClass = namedtuple(
@@ -35,50 +32,15 @@ Class3DParticleClass.overall_fourier_completeness.__doc__ = (
 )
 
 
-class Class3D(collections.abc.Mapping):
-    def __eq__(self, other):
-        if isinstance(other, Class3D):
-            return self._basepath == other._basepath
-        return False
-
+class Class3D(JobType):
     def __hash__(self):
         return hash(("relion._parser.Class3D", self._basepath))
-
-    def __init__(self, path):
-        self._basepath = path
-        self._jobcache = {}
-
-    def __iter__(self):
-        return iter(self.jobs)
-
-    def __len__(self):
-        return len(self.jobs)
 
     def __repr__(self):
         return f"Class3D({repr(str(self._basepath))})"
 
     def __str__(self):
         return f"<Class3D parser at {self._basepath}>"
-
-    @property
-    def jobs(self):
-        return sorted(
-            d.stem
-            for d in self._basepath.iterdir()
-            if d.is_dir() and not d.is_symlink()
-        )
-
-    def __getitem__(self, key):
-        if not isinstance(key, str):
-            raise KeyError(f"Invalid argument {key!r}, expected string")
-        if key not in self._jobcache:
-            job_path = self._basepath / key
-            if not job_path.is_dir():
-                raise KeyError(
-                    f"no job directory present for {key} in {self._basepath}"
-                )
-            self._jobcache[key] = self._load_job_directory(key)
-        return self._jobcache[key]
 
     @property
     def job_number(self):
@@ -140,21 +102,6 @@ class Class3D(collections.abc.Mapping):
             if not check_file.exists():
                 raise ValueError(f"File {check_file} missing from job directory")
         return data_file, model_file
-
-    @functools.lru_cache(maxsize=None)
-    def _read_star_file(self, job_num, file_name):
-        full_path = self._basepath / job_num / file_name
-        gemmi_readable_path = os.fspath(full_path)
-        star_doc = cif.read_file(gemmi_readable_path)
-        return star_doc
-
-    def parse_star_file(self, loop_name, star_doc, block_number):
-        data_block = star_doc[block_number]
-        values = data_block.find_loop(loop_name)
-        values_list = list(values)
-        if not values_list:
-            print("Warning - no values found for", loop_name)
-        return values_list
 
     def _class_checker(
         self, tuple_list, length
