+++ src/relion/_parser/ctffind.py
@@ -1,8 +1,5 @@
-import collections.abc
-from gemmi import cif
-import os
-import functools
 from collections import namedtuple
+from relion._parser.jobtype import JobType
 
 CTFMicrograph = namedtuple(
     "CTFMicrograph",
@@ -34,24 +31,9 @@ CTFMicrograph.fig_of_merit.__doc__ = (
 )
 
 
-class CTFFind(collections.abc.Mapping):
-    def __eq__(self, other):
-        if isinstance(other, CTFFind):
-            return self._basepath == other._basepath
-        return False
-
+class CTFFind(JobType):
     def __hash__(self):
         return hash(("relion._parser.CTFFind", self._basepath))
-
-    def __init__(self, path):
-        self._basepath = path
-        self._jobcache = {}
-
-    def __iter__(self):
-        return iter(self.jobs)
-
-    def __len__(self):
-        return len(self.jobs)
 
     def __repr__(self):
         return f"CTFFind({repr(str(self._basepath))})"
@@ -60,32 +42,12 @@ class CTFFind(collections.abc.Mapping):
         return f"<CTFFind parser at {self._basepath}>"
 
     @property
-    def jobs(self):
-        return sorted(
-            d.stem
-            for d in self._basepath.iterdir()
-            if d.is_dir() and not d.is_symlink()
-        )
-
-    def __getitem__(self, key):
-        if not isinstance(key, str):
-            raise KeyError(f"Invalid argument {key!r}, expected string")
-        if key not in self._jobcache:
-            job_path = self._basepath / key
-            if not job_path.is_dir():
-                raise KeyError(
-                    f"no job directory present for {key} in {self._basepath}"
-                )
-            self._jobcache[key] = self._load_job_directory(key)
-        return self._jobcache[key]
-
-    @property
     def job_number(self):
         jobs = [x.name for x in self._basepath.iterdir()]
         return jobs
 
     def _load_job_directory(self, jobdir):
-        file = self._read_star_file(jobdir)
+        file = self._read_star_file(jobdir, "micrographs_ctf.star")
 
         astigmatism = self.parse_star_file("_rlnCtfAstigmatism", file, 1)
         defocus_u = self.parse_star_file("_rlnDefocusU", file, 1)
@@ -110,18 +72,3 @@ class CTFFind(collections.abc.Mapping):
                 )
             )
         return micrograph_list
-
-    def parse_star_file(self, loop_name, star_doc, block_number):
-        data_block = star_doc[block_number]
-        values = data_block.find_loop(loop_name)
-        values_list = list(values)
-        if not values_list:
-            print("Warning - no values found for", loop_name)
-        return values_list
-
-    @functools.lru_cache(maxsize=None)
-    def _read_star_file(self, job_num):
-        full_path = self._basepath / job_num / "micrographs_ctf.star"
-        gemmi_readable_path = os.fspath(full_path)
-        star_doc = cif.read_file(gemmi_readable_path)
-        return star_doc
